Extract Decision dataclass
