+++ strategies/__init__.py
@@ -1,4 +1,5 @@
-from .base import BaseStrategy, Decision
+from core_strategy import Decision
+from .base import BaseStrategy
 from .momentum import MomentumStrategy
 
 __all__ = [
