+++ strategies/base.py
@@ -1,50 +1,9 @@
 # strategies/base.py
 from __future__ import annotations
 
-from dataclasses import dataclass
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List
 
-from core_strategy import Strategy
-
-from action_proto import ActionProto, ActionType
-
-
-@dataclass(frozen=True)
-class Decision:
-    """
-    Результат решения стратегии на шаге.
-
-    Поля:
-      - side: "BUY" | "SELL"
-      - volume_frac: целевая величина заявки в долях позиции ([-1.0; 1.0]).
-        Для MARKET это величина, знак задаёт сторону.
-      - price_offset_ticks: сдвиг цены в тиках для LIMIT (по умолчанию 0; для MARKET игнорируется)
-      - tif: GTC|IOC|FOK (строка, совместимо с ActionProto.tif)
-      - client_tag: произвольная метка стратегии
-    """
-    side: str
-    volume_frac: float
-    price_offset_ticks: int = 0
-    tif: str = "GTC"
-    client_tag: Optional[str] = None
-
-    def to_action_proto(self) -> ActionProto:
-        """
-        Преобразование решения в ActionProto без потери информации.
-        Для side="BUY" volume_frac >= 0; для "SELL" — volume_frac <= 0.
-        """
-        v = float(self.volume_frac)
-        if str(self.side).upper() == "SELL":
-            v = -abs(v)
-        else:
-            v = abs(v)
-        return ActionProto(
-            action_type=(ActionType.MARKET if self.price_offset_ticks == 0 else ActionType.LIMIT),
-            volume_frac=v,
-            price_offset_ticks=int(self.price_offset_ticks),
-            tif=str(self.tif),
-            client_tag=self.client_tag,
-        )
+from core_strategy import Decision, Strategy
 
 
 class BaseStrategy(Strategy):
@@ -82,3 +41,6 @@ class BaseStrategy(Strategy):
         Реализации должны вернуть список Decision.
         """
         return []
+
+
+__all__ = ["BaseStrategy", "Decision"]
